src/auth/utils.py
import bcrypt
import jwt
from src.config import settings


def encode_jwt(
    payload: dict,
    private_key: str = settings.private_key_path.read_text(),
    algorithm: str = settings.algorithm,
) -> str:
    return jwt.encode(payload, private_key, algorithm)


def decode_jwt(
    token: str | bytes,
    public_key: str = settings.public_key_path.read_text(),
    algorithm: str = settings.algorithm,
) -> dict | None:
    try:
        return jwt.decode(token, public_key, algorithms=[algorithm])
    except jwt.DecodeError:
        return None


def hash_password(password: str) -> str:
    salt = bcrypt.gensalt()
    password_bytes = password.encode()
    return bcrypt.hashpw(password_bytes, salt).decode("utf-8")


def validate_password(password: str, hashed_password: str) -> bool:
    return bcrypt.checkpw(password.encode(), hashed_password.encode("utf-8"))
