requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
alembic==1.13.1
    # via -r requirements.in
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via
    #   starlette
    #   watchfiles
asyncpg==0.29.0
    # via -r requirements.in
bcrypt==4.1.2
    # via -r requirements.in
cffi==1.16.0
    # via cryptography
click==8.1.7
    # via uvicorn
cryptography==42.0.5
    # via pyjwt
dnspython==2.6.1
    # via email-validator
email-validator==2.1.1
    # via pydantic
fastapi==0.110.0
    # via -r requirements.in
greenlet==3.0.3
    # via sqlalchemy
h11==0.14.0
    # via uvicorn
httptools==0.6.1
    # via uvicorn
idna==3.6
    # via
    #   anyio
    #   email-validator
mako==1.3.2
    # via alembic
markupsafe==2.1.5
    # via mako
pycparser==2.21
    # via cffi
pydantic[email]==2.6.3
    # via
    #   -r requirements.in
    #   fastapi
    #   pydantic-settings
pydantic-core==2.16.3
    # via pydantic
pydantic-settings==2.2.1
    # via -r requirements.in
pyjwt[crypto]==2.8.0
    # via -r requirements.in
python-dotenv==1.0.1
    # via
    #   pydantic-settings
    #   uvicorn
pyyaml==6.0.1
    # via uvicorn
sniffio==1.3.1
    # via anyio
sqlalchemy[asyncio]==2.0.27
    # via
    #   -r requirements.in
    #   alembic
starlette==0.36.3
    # via fastapi
typing-extensions==4.10.0
    # via
    #   alembic
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
uvicorn[standard]==0.27.1
    # via -r requirements.in
uvloop==0.19.0
    # via uvicorn
watchfiles==0.21.0
    # via uvicorn
websockets==12.0
    # via uvicorn
